+++ opentelemetry-zipkin/examples/zipkin.rs
@@ -1,7 +1,10 @@
 use opentelemetry::{
     global::{self},
-    trace::{Span, Tracer},
+    trace::{Span, TraceError, Tracer},
+    InstrumentationScope, KeyValue,
 };
+use opentelemetry_sdk::trace::TracerProvider;
+use opentelemetry_zipkin::ZipkinExporter;
 use std::thread;
 use std::time::Duration;
 
@@ -12,10 +15,26 @@ fn bar() {
     span.end()
 }
 
-fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
-    let (tracer, provider) = opentelemetry_zipkin::new_pipeline()
+fn init_traces() -> Result<TracerProvider, TraceError> {
+    let exporter = ZipkinExporter::builder()
         .with_service_name("trace-demo")
-        .install_simple()?;
+        .build()?;
+
+    Ok(TracerProvider::builder()
+        .with_simple_exporter(exporter)
+        .build())
+}
+
+fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
+    let provider = init_traces()?;
+    global::set_tracer_provider(provider.clone());
+
+    let common_scope_attributes = vec![KeyValue::new("scope-key", "scope-value")];
+    let scope = InstrumentationScope::builder("opentelemetry-zipkin")
+        .with_version(env!("CARGO_PKG_VERSION"))
+        .with_attributes(common_scope_attributes)
+        .build();
+    let tracer = global::tracer_with_scope(scope.clone());
 
     tracer.in_span("foo", |_cx| {
         thread::sleep(Duration::from_millis(6));